build speedup: Do not use shell 'find' function to look for NOTICE files.
The GNU Make wildcard function can be used to look for a specific file and
is _much_ faster than invoking the find command through $(shell ...).

Since notice_files.mk is included by base_rules.mk for each and every module
in the build tree, this represents a considerable speed-up. For example, when
building the full-eng product from the AOSP tree, the no-op "make" goes from
21 seconds to 16 seconds with this patch applied.

Change-Id: Ideb1bb74d5bfbb650834fa4b6e151ff24dcd0150